# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.knowledgeItem:
    properties:
      children:
        items:
          $ref: '#/definitions/main.knowledgeItem'
        type: array
      confidence:
        type: number
      content:
        type: string
      id:
        type: string
      itemRole:
        type: string
      itemType:
        type: string
      sourceMessageIds:
        items:
          type: string
        type: array
      sourceMessages:
        items:
          $ref: '#/definitions/main.knowledgeSourceMessage'
        type: array
      updatedAt:
        type: string
    type: object
  main.knowledgeRebuildJobView:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      failedCount:
        type: integer
      id:
        type: string
      processedCount:
        type: integer
      replaceExisting:
        type: boolean
      skippedCount:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      totalCount:
        type: integer
      updatedAt:
        type: string
    type: object
  main.knowledgeSourceMessage:
    properties:
      documentSummary:
        type: string
      groupId:
        type: string
      hasMedia:
        type: boolean
      id:
        type: string
      imageUrl:
        type: string
      kind:
        type: string
      mediaMime:
        type: string
      mediaStatus:
        type: string
      mediaUrl:
        type: string
      receivedAt:
        type: string
      senderJid:
        type: string
      senderName:
        type: string
      text:
        type: string
      thumbnailUrl:
        type: string
    type: object
  main.knowledgeTopic:
    properties:
      confidence:
        type: number
      groupId:
        type: string
      groupSubject:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/main.knowledgeItem'
        type: array
      parentTopicId:
        type: string
      sourceMessageIds:
        items:
          type: string
        type: array
      subtopicKey:
        type: string
      subtopics:
        items:
          $ref: '#/definitions/main.knowledgeTopic'
        type: array
      summary:
        type: string
      title:
        type: string
      topicKey:
        type: string
      updatedAt:
        type: string
    type: object
  main.knowledgeTopicDefinitionPage:
    properties:
      items:
        items:
          $ref: '#/definitions/main.knowledgeTopicDefinitionView'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  main.knowledgeTopicDefinitionRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      language:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
      topicKey:
        type: string
    type: object
  main.knowledgeTopicDefinitionView:
    properties:
      createdAt:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      language:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
      topicKey:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact:
    name: CONXTOR project
    url: https://github.com/conxtor/groupmaster
  description: |-
    REST API for ingesting, analyzing, and presenting selected WhatsApp and Telegram group conversations.
    The public deployment serves this API below the /api reverse-proxy prefix.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/conxtor/groupmaster
  title: CONXTOR Messaging Group Intelligence API
  version: 0.1.0
paths:
...
  /admin/knowledge/topics:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Language code
        in: query
        name: language
        type: string
      - description: Search topic key, title, or description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.knowledgeTopicDefinitionPage'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List Knowledge Base topic definitions
      tags:
      - administration
    post:
      consumes:
      - application/json
      parameters:
      - description: Topic definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.knowledgeTopicDefinitionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.knowledgeTopicDefinitionView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create Knowledge Base topic definition
      tags:
      - administration
  /admin/knowledge/topics/{id}:
    delete:
      parameters:
      - description: Topic definition UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete Knowledge Base topic definition
      tags:
      - administration
    patch:
      consumes:
      - application/json
      parameters:
      - description: Topic definition UUID
        in: path
        name: id
        required: true
        type: string
      - description: Topic definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.knowledgeTopicDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.knowledgeTopicDefinitionView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update Knowledge Base topic definition
      tags:
      - administration
  /admin/knowledge/topics/rebuild:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.knowledgeRebuildJobView'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List Knowledge Base rebuild jobs
      tags:
      - administration
    post:
      consumes:
      - application/json
      parameters:
      - description: Rebuild options
        in: body
        name: body
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.knowledgeRebuildJobView'
            type: array
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.knowledgeRebuildJobView'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Rebuild Knowledge Base
      tags:
      - administration
...
  /knowledge:
    get:
      parameters:
      - description: Filter by group ID
        in: query
        name: groupId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.knowledgeTopic'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List Knowledge Base topics
      tags:
      - knowledge
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: Authenticated CONXTOR session cookie.
    in: cookie
    name: wagi_session
    type: apiKey
swagger: "2.0"